further quantizer optimization
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@4714 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ base/Profiler.C
@@ -76,7 +76,7 @@ void Profiles::dump()
 
         cerr << "-> " << *i << ": " 
 	     << m_profiles[*i].first << " calls, "
-	     << ((m_profiles[*i].second * 1000) / CLOCKS_PER_SEC) << "ms, "
+	     << int((m_profiles[*i].second * 1000.0) / CLOCKS_PER_SEC) << "ms, "
 	     << (((double)m_profiles[*i].second * 1000000.0 /
 		  (double)m_profiles[*i].first) / CLOCKS_PER_SEC) << "us/call"
 	     << endl;
